+++ front-end/src/app/components/benevoles/single-benevole-affectations/single-benevole-affectations.component.ts
@@ -15,6 +15,8 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 // Error Dialog
 import { ErrorDialogComponent } from '../../partials/error-dialog/error-dialog.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
+// auth
+import { AuthService } from 'src/app/services/auth.service';
 
 @Component({
   selector: 'festivalJeux-single-benevole-affectations',
@@ -32,6 +34,10 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
 
   valid: any = {};
 
+  // auth
+  userId! : any
+  isAuth! : boolean
+
   // zones data
   zones: Zone[] = [];
   zonesSub!: Subscription;
@@ -47,10 +53,12 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
     private route: ActivatedRoute,
     private router :  Router,
     private snackBar: MatSnackBar,
+    private authService: AuthService,
   ) { }
 
   ngOnInit(): void {
-
+    // auth
+    this.verifSignIn()
     //Get benevole
     // récuperer l'id depuis la route
     this.route.params.subscribe({
@@ -82,6 +90,19 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
       }
     })
     this.zonesService.getZones()
+  }
+
+  //assigner is auth a true si user est connecté
+  verifSignIn() : void{
+    this.authService.isAuth$.subscribe(
+      (bool: boolean)=>{
+        this.isAuth = bool
+        if(bool){
+          // get user id
+          this.userId = this.authService.userId
+        }
+      }
+    )
   }
 
   // init benevole subscription
@@ -119,20 +140,31 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
   }
 
   addRow() {
-    this.rowNumber--;
-    this.valid[this.rowNumber] = {
-      prenom : false,
-      nom: false,
-      email: false,
-      zone: false,
-      debut_creneau: false,
-      fin_creneau: false,
-      date: false,
-    }
-    // create the same BenevoleDisplay object as the one in the table but with empty values for the affectation
-    let benevoleDisplayRow = new BenevoleDisplay(this.benevole, new Affectation(new Zone("",""), new Creneau("", "", "")));
-    benevoleDisplayRow.isEdit = true;
-    this.benevolesDisplay.data = [benevoleDisplayRow, ...this.benevolesDisplay.data];
+    // if the user is not connected, he can't add a row
+    if(this.isAuth){
+      this.rowNumber--;
+      this.valid[this.rowNumber] = {
+        prenom : false,
+        nom: false,
+        email: false,
+        zone: false,
+        debut_creneau: false,
+        fin_creneau: false,
+        date: false,
+      }
+      // create the same BenevoleDisplay object as the one in the table but with empty values for the affectation
+      let benevoleDisplayRow = new BenevoleDisplay(this.benevole, new Affectation(new Zone("",""), new Creneau("", "", "")));
+      benevoleDisplayRow.isEdit = true;
+      this.benevolesDisplay.data = [benevoleDisplayRow, ...this.benevolesDisplay.data];
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+      data: 'You must be connected to add a row',
+      duration: 3000,
+      panelClass: ['error-snackbar'],
+    });
+    throw new Error('You must be connected to add a row');
+  }
+
   }
 
   /**
@@ -143,9 +175,19 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
    *
    */
   editRow(row: BenevoleDisplay){
-    let benevole = this.createBenevoleFromDisplay(row);
-    this.updateAffectations(row,benevole.affectations)
+    // if the user is not connected, he can't edit a row
+    if(this.isAuth){
+      let benevole = this.createBenevoleFromDisplay(row);
+      this.updateAffectations(row,benevole.affectations)
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+      data: 'You must be connected to edit a row',
+      duration: 3000,
+      panelClass: ['error-snackbar'],
+    });
+    throw new Error('You must be connected to edit a row');
   }
+}
 
   /**
  * Cancel the edit of a row
@@ -172,14 +214,24 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
    *
   */
   createBenevoleFromDisplay(row: BenevoleDisplay): Benevole{
-    let affectations: Affectation[] = [];
-    this.benevolesDisplay.data.forEach((benevoleDisplay: BenevoleDisplay) => {
-      let affectation = this.createAffectationFromBenevoleDisplay(benevoleDisplay)
-      affectations.push(affectation);
-    });
-    let benevole = new Benevole(row._id,row.prenom, row.nom, row.email, affectations);
-    return benevole;
-  }
+    // if the user is not connected, he can't edit a row
+    if(this.isAuth){
+      let affectations: Affectation[] = [];
+      this.benevolesDisplay.data.forEach((benevoleDisplay: BenevoleDisplay) => {
+        let affectation = this.createAffectationFromBenevoleDisplay(benevoleDisplay)
+        affectations.push(affectation);
+      });
+      let benevole = new Benevole(row._id,row.prenom, row.nom, row.email, affectations);
+      return benevole;
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'You must be connected to add a row',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
+      });
+      throw new Error('You must be connected to add a row');
+    }
+}
 
   /**
    * createAffectationFromBenevoleDisplay
@@ -187,34 +239,78 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
    * @returns Affectation
   */
   createAffectationFromBenevoleDisplay(row: BenevoleDisplay): Affectation {
+    // if the user is not connected, he can't edit a row
+    if(this.isAuth){
     const debutCreneau = new Date(row.date + " " + row.debut_creneau);
     const finCreneau = new Date(row.date + " " + row.fin_creneau);
 
+    // check if the end time is after the start time
     if (finCreneau.getTime() <= debutCreneau.getTime()) {
       this.snackBar.openFromComponent(ErrorDialogComponent, {
         data: 'End time must be after start time',
         duration: 3000,
+        panelClass: ['error-snackbar'],
       });
       throw new Error('End time must be after start time');
     }
 
+    // Check if for this benevole the interval debutCreneau / finCreneau does not intersect with another interval
+    // complete de args for the some function to check if the interval is already taken
+
+    /* const intersect = this.benevolesDisplay.data.some(function (benevoleDisplay: BenevoleDisplay) {
+      if (benevoleDisplay._id !== row._id) {
+        const debutCreneauBenevole = new Date(benevoleDisplay.date + " " + benevoleDisplay.debut_creneau);
+        const finCreneauBenevole = new Date(benevoleDisplay.date + " " + benevoleDisplay.fin_creneau);
+        return (
+          (debutCreneau.getTime() >= debutCreneauBenevole.getTime() && debutCreneau.getTime() < finCreneauBenevole.getTime()) ||
+          (finCreneau.getTime() > debutCreneauBenevole.getTime() && finCreneau.getTime() <= finCreneauBenevole.getTime())
+        );
+      }
+    });
+    if (intersect) {
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'The interval is already taken',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
+      });
+      throw new Error('The interval is already taken');
+    } */
+
+
     const zone: Zone = this.zones.find((zone: Zone) => zone._id === row.idZone) || new Zone('', '');
     const creneau: Creneau = new Creneau(row.debut_creneau, row.fin_creneau, row.date);
     return new Affectation(zone, creneau);
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'You must be connected to add a row',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
+      });
+      throw new Error('You must be connected to add a row');
+    }
 
   }
 
   removeRow(row: BenevoleDisplay) {
-    this.dialog
-      .open(ConfirmDialogComponent)
-      .afterClosed()
-      .subscribe((confirm) => {
-        if (confirm) {
-          // create an affectation object from the row
-          let affectation = this.createAffectationFromBenevoleDisplay(row)
-          this.deleteAffectation(row._id,affectation)
-        }
+    // if the user is not connected, he can't edit a row
+    if(this.isAuth){
+      this.dialog
+        .open(ConfirmDialogComponent)
+        .afterClosed()
+        .subscribe((confirm) => {
+          if (confirm) {
+            // create an affectation object from the row
+            let affectation = this.createAffectationFromBenevoleDisplay(row)
+            this.deleteAffectation(row._id,affectation)
+          }
+        });
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'You must be connected to add a row',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
       });
+    }
   }
 
   isAllSelected() {
@@ -240,20 +336,29 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
    * else it will do nothing
    */
   removeSelectedRows() {
-    this.dialog
-      .open(ConfirmDialogComponent)
-      .afterClosed()
-      .subscribe((confirm) => {
-        if (confirm) {
-          this.benevolesDisplay.data.forEach((row: BenevoleDisplay) => {
-            if (row.isSelected) {
-              // create an affectation object from the row
-              let affectation = this.createAffectationFromBenevoleDisplay(row)
-              this.deleteAffectation(row._id,affectation)
-            }
-          });
-        }
+    // if the user is not connected, he can't edit a row
+    if(!this.isAuth){
+      this.dialog
+        .open(ConfirmDialogComponent)
+        .afterClosed()
+        .subscribe((confirm) => {
+          if (confirm) {
+            this.benevolesDisplay.data.forEach((row: BenevoleDisplay) => {
+              if (row.isSelected) {
+                // create an affectation object from the row
+                let affectation = this.createAffectationFromBenevoleDisplay(row)
+                this.deleteAffectation(row._id,affectation)
+              }
+            });
+          }
+        });
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'You must be connected to add a row',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
       });
+    }
   }
 
   /**
@@ -265,13 +370,22 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
    * else it will display an error message
   */
   updateAffectations(row: BenevoleDisplay, affectations: Affectation[]) {
-    this.benevolesService.updateAffectations(row._id, affectations).then(()=>{
-      // change the isEdit property to false to stop the edition mode
-      row.isEdit = false
-    })
-    .catch((err)=>{
-      console.log(err.message)
-    })
+    // if the user is not connected, he can't edit a row
+    if(!this.isAuth){
+      this.benevolesService.updateAffectations(row._id, affectations).then(()=>{
+        // change the isEdit property to false to stop the edition mode
+        row.isEdit = false
+      })
+      .catch((err)=>{
+        console.log(err.message)
+      })
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'You must be connected to add a row',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
+      });
+    }
   }
 
   /**
@@ -282,12 +396,21 @@ export class SingleBenevoleAffectationsComponent implements OnInit {
    * else it will display an error message
   */
   deleteAffectation(_id: string, affectation: Affectation) {
-    this.benevolesService.deleteAffectation(_id, affectation).then(()=>{
-        this.benevolesDisplay.data = this.benevolesDisplay.data.filter((u) => u._id !== _id);
-    })
-    .catch((err)=>{
-      console.log(err.message)
-    })
+    // if the user is not connected, he can't edit a row
+    if(!this.isAuth){
+      this.benevolesService.deleteAffectation(_id, affectation).then(()=>{
+          this.benevolesDisplay.data = this.benevolesDisplay.data.filter((u) => u._id !== _id);
+      })
+      .catch((err)=>{
+        console.log(err.message)
+      })
+    }else{
+      this.snackBar.openFromComponent(ErrorDialogComponent, {
+        data: 'You must be connected to add a row',
+        duration: 3000,
+        panelClass: ['error-snackbar'],
+      });
+    }
   }
